Continuing App Design
this code is our first rough draft focused on the snake's movement and directions as well as starting to add the food/points. this also coded for two of the error statements that could occur. this was worked on during class, seperately, and a group meet up. as a group, we also worked on researching and compiling code for game obstacles/easter eggs on a document to add in later.