Commenting on Code
After working with Nash and Priyanka on the code during class and after (during our meeting), I added comments to the code to explain the functionality of the movement. The purpose of this is because this code is like a "rough draft," so the comments will allow to be more efficient when debugging the code in the future.