Debugging Boundaries of the Game
After class, Priyanka, Nash, and I spent an hour trying to figure out how to get the boundaries of the snake to work. Initially, the error message that the snake was out of bounds would show up when the snake was still in the middle of the plot. We fixed this so that the error message would only appear when the snake was truly hitting the boundaries.